Add files via upload
Version 2.0.0.0  is the second release of PH330Lib and primarily provided
to support the new PicoHarp 330 4P with four input channels.
The previous model with two input channels is of course also supported.
A very useful new feature is the API call PH330_SaveDebugDump. It is
provided to help debugging gateware issues by letting the user save a
snapshot of the device’s internal FPGA states to a file that then
can be submitted for support.